+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "ultimate64-manager"
 version = "0.3.21"
-edition = "2024"
+edition = "2021"
 description = "Ultimate64 Manager - Control your Ultimate64/Ultimate-II+"
 authors = ["Marcin Spoczynski"]
 license = "MIT"
