+++ src/api.rs
@@ -1,5 +1,6 @@
 // Ultimate64 REST API client
 
+use crate::net_utils::REST_TIMEOUT_SECS;
 use reqwest::Client;
 use std::sync::Arc;
 use tokio::sync::Mutex;
@@ -191,5 +192,113 @@ pub async fn run_disk(
             "Mounted: {} (no connection for auto-run)",
             filename
         ))
+    }
+}
+
+// ─────────────────────────────────────────────────────────────────
+//  Memory read/write operations (via ultimate64 crate)
+// ─────────────────────────────────────────────────────────────────
+
+/// Maximum bytes per REST write chunk (mirrors the C++ SOCKET_BUFFER_SIZE guard).
+const RAW_CHUNK: usize = 256;
+
+pub async fn read_memory_async(
+    connection: Arc<Mutex<Rest>>,
+    address: u16,
+    length: u16,
+) -> Result<Vec<u8>, String> {
+    let result = tokio::time::timeout(
+        tokio::time::Duration::from_secs(REST_TIMEOUT_SECS),
+        tokio::task::spawn_blocking(move || {
+            let conn = connection.blocking_lock();
+            conn.read_mem(address, length)
+                .map_err(|e| format!("Read failed: {}", e))
+        }),
+    )
+    .await;
+    match result {
+        Ok(Ok(data)) => data,
+        Ok(Err(e)) => Err(format!("Task error: {}", e)),
+        Err(_) => Err("Read timed out — device may be offline".to_string()),
+    }
+}
+
+pub async fn write_byte_async(
+    connection: Arc<Mutex<Rest>>,
+    address: u16,
+    value: u8,
+) -> Result<(), String> {
+    let result = tokio::time::timeout(
+        tokio::time::Duration::from_secs(REST_TIMEOUT_SECS),
+        tokio::task::spawn_blocking(move || {
+            let conn = connection.blocking_lock();
+            conn.write_mem(address, &[value])
+                .map_err(|e| format!("Write failed: {}", e))
+        }),
+    )
+    .await;
+    match result {
+        Ok(Ok(r)) => r,
+        Ok(Err(e)) => Err(format!("Task error: {}", e)),
+        Err(_) => Err("Write timed out — device may be offline".to_string()),
+    }
+}
+
+pub async fn fill_memory_async(
+    connection: Arc<Mutex<Rest>>,
+    address: u16,
+    length: u16,
+    value: u8,
+) -> Result<(), String> {
+    let result = tokio::time::timeout(
+        tokio::time::Duration::from_secs(REST_TIMEOUT_SECS * 2),
+        tokio::task::spawn_blocking(move || {
+            let conn = connection.blocking_lock();
+            let fill_data: Vec<u8> = vec![value; RAW_CHUNK];
+            let mut offset = 0u16;
+            while offset < length {
+                let remaining = (length - offset) as usize;
+                let write_size = remaining.min(RAW_CHUNK);
+                let current_addr = address.wrapping_add(offset);
+                conn.write_mem(current_addr, &fill_data[..write_size])
+                    .map_err(|e| format!("Fill failed at ${:04X}: {}", current_addr, e))?;
+                offset += write_size as u16;
+            }
+            Ok(())
+        }),
+    )
+    .await;
+    match result {
+        Ok(Ok(r)) => r,
+        Ok(Err(e)) => Err(format!("Task error: {}", e)),
+        Err(_) => Err("Fill timed out — device may be offline".to_string()),
+    }
+}
+
+pub async fn write_memory_async(
+    connection: Arc<Mutex<Rest>>,
+    address: u16,
+    data: Vec<u8>,
+) -> Result<(), String> {
+    let result = tokio::time::timeout(
+        tokio::time::Duration::from_secs(REST_TIMEOUT_SECS * 4),
+        tokio::task::spawn_blocking(move || {
+            let conn = connection.blocking_lock();
+            let mut offset = 0usize;
+            while offset < data.len() {
+                let write_size = (data.len() - offset).min(RAW_CHUNK);
+                let current_addr = address.wrapping_add(offset as u16);
+                conn.write_mem(current_addr, &data[offset..offset + write_size])
+                    .map_err(|e| format!("Write failed at ${:04X}: {}", current_addr, e))?;
+                offset += write_size;
+            }
+            Ok(())
+        }),
+    )
+    .await;
+    match result {
+        Ok(Ok(r)) => r,
+        Ok(Err(e)) => Err(format!("Task error: {}", e)),
+        Err(_) => Err("Write timed out — device may be offline".to_string()),
     }
 }
